refactor: #15 fix Status enum import paths to use global.entity package
