+++ src/test/java/laughcandidate/yellowribbonbe/admin/controller/AdminBadgeControllerTest.java
@@ -6,7 +6,7 @@ import laughcandidate.yellowribbonbe.admin.dto.response.BadgeApplyListResponse;
 import laughcandidate.yellowribbonbe.admin.service.AdminBadgeService;
 import laughcandidate.yellowribbonbe.badge.entity.BadgeApply;
 import laughcandidate.yellowribbonbe.badge.entity.Category;
-import laughcandidate.yellowribbonbe.badge.entity.Status;
+import laughcandidate.yellowribbonbe.global.entity.Status;
 import laughcandidate.yellowribbonbe.global.exception.CustomException;
 import laughcandidate.yellowribbonbe.global.exception.errorCode.AdminErrorCode;
 import org.junit.jupiter.api.DisplayName;
